# ChunkCache: reduce the overhead of cache hits while updating metrics