The intent here is to reduce the overhead in updating the ChunkCache metrics in case of cache hit.

![image](https://github.com/datastax/cassandra/assets/9469110/54226848-055b-4799-aa26-f119cb7dd6e6)