@eolivelli 2024-03-20
I did some tests on a reference workload with SAI and NSS on Fallout

It seems that this patch doesn't improve the latency consistently.
From the flamegraphs I do see that the  numbers of calls to "increment" are fewer, and that patch has some effect, but it is not a big deal.

WITH THIS PATCH

http://fallout.aws.dsinternal.org/tests/ui/enrico.olivelli@datastax.com/NSSIndexTests/4ac2cb7c-8c28-4078-89e7-451e974b6354/artifacts/server/performance-report.html?metric=Ops_Sec&operation=reads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation=secondreads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation_filter=.%2Bsuccess&smoothing=1&show_aggregates=true&min_ops=0&xmin=0&xmax=1111.02&ymin=0&ymax=1209.78

Test Duration              : 16m 46s
Total Operations           : 1000000
Op Rate                    : 993 op/sec
Min Latency                : 2.696 ms
Avg Latency                : 3.369 ms
Median Latency             : 3.153 ms
95th Latency               : 4.510 ms
99th Latency               : 6.107 ms
99.9th Latency             : 8.174 ms
Max Latency                : 32.804 ms
Median Absolute Deviation  : 0.133 ms
Interquartile Range        : 0.332 ms


http://fallout.aws.dsinternal.org/tests/ui/enrico.olivelli@datastax.com/NSSIndexTests/b5d8c461-b630-4339-9039-4e368ec41be7/artifacts/server/performance-report.html?metric=Ops_Sec&operation=reads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation=secondreads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation_filter=.%2Bsuccess&smoothing=1&show_aggregates=true&min_ops=0&xmin=0&xmax=1111.01&ymin=0&ymax=1205.93



Test Duration              : 16m 46s
Total Operations           : 1000000
Op Rate                    : 993 op/sec
Min Latency                : 2.114 ms
Avg Latency                : 2.570 ms
Median Latency             : 2.444 ms
95th Latency               : 3.299 ms
99th Latency               : 4.179 ms
99.9th Latency             : 6.105 ms
Max Latency                : 33.439 ms
Median Absolute Deviation  : 0.078 ms
Interquartile Range        : 0.167 ms


BEFORE THE PATCH

http://fallout.aws.dsinternal.org/tests/ui/enrico.olivelli@datastax.com/NSSIndexTests/147312ca-2224-432e-8db8-ef0072802366/artifacts/server/performance-report.html?metric=Ops_Sec&operation=reads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation=secondreads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation_filter=.%2Bsuccess&smoothing=1&show_aggregates=true&min_ops=0&xmin=0&xmax=1111.01&ymin=0&ymax=1227.6

Test Duration              : 16m 46s
Total Operations           : 1000000
Op Rate                    : 993 op/sec
Min Latency                : 3.891 ms
Avg Latency                : 4.966 ms
Median Latency             : 4.687 ms
95th Latency               : 6.665 ms
99th Latency               : 7.188 ms
99.9th Latency             : 9.168 ms
Max Latency                : 406.847 ms




http://fallout.aws.dsinternal.org/tests/ui/enrico.olivelli@datastax.com/NSSIndexTests/8376d920-2f2b-4dc6-8671-92ef3a996b04/artifacts/server/performance-report.html?metric=Ops_Sec&operation=reads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation=secondreads_9dfd7719-b7c2-43b4-853e-f8604576d5db_enrico-nss-18m-sharedcc%3Aresult_success&operation_filter=.%2Bsuccess&smoothing=1&show_aggregates=true&min_ops=0&xmin=0&xmax=1111.01&ymin=0&ymax=1202.96

---------------------------------------
Test Duration              : 16m 46s
Total Operations           : 1000000
Op Rate                    : 993 op/sec
Min Latency                : 2.084 ms
Avg Latency                : 2.626 ms
Median Latency             : 2.568 ms
95th Latency               : 3.183 ms
99th Latency               : 3.658 ms
99.9th Latency             : 4.821 ms
Max Latency                : 49.566 ms
Median Absolute Deviation  : 0.103 ms
Interquartile Range        : 0.212 ms