@eolivelli 2024-03-20
I am going to try with a Vector workload, in which we do more accesses to the ChunkCache